Initial creation of Data.DList.Unsafe
