+++ dlist.cabal
@@ -40,6 +40,7 @@ library
                         deepseq >= 1.1 && < 1.5
   extensions:           CPP
   exposed-modules:      Data.DList
+                        Data.DList.Unsafe
   ghc-options:          -Wall
 
 test-suite test
